+++ src/module/applications/apps/_module.mjs
@@ -1,2 +1,4 @@
 export { default as PowerRollDialog } from "./power-roll-dialog.mjs";
+export { default as SavingThrowDialog } from "./saving-throw-dialog.mjs";
+export { default as SavingThrowManager } from "./saving-throw-manager.mjs";
 export { default as TargetedConditionPrompt } from "./targeted-condition-prompt.mjs";