+++ src/module/documents/combat.mjs
@@ -1,8 +1,11 @@
+import SavingThrowManager from "../applications/apps/saving-throw-manager.mjs";
 import { systemID } from "../constants.mjs";
+import BaseEffectModel from "../data/effect/base.mjs";
 import { DSRoll } from "../rolls/base.mjs";
 
+/** @import ActiveEffectData from "@common/documents/_types.mjs" */
 /** @import { MaliceModel } from "../data/settings/malice.mjs" */
-/** @import { DrawSteelCombatant, DrawSteelCombatantGroup } from "./_module.mjs" */
+/** @import { DrawSteelActor, DrawSteelCombatant, DrawSteelCombatantGroup } from "./_module.mjs" */
 
 /**
  * A document subclass adding system-specific behavior and registered in CONFIG.Combat.documentClass
@@ -13,6 +16,8 @@ export default class DrawSteelCombat extends foundry.documents.Combat {
     super.prepareDerivedData();
     Hooks.callAll("ds.prepareCombatData", this);
   }
+
+  /* -------------------------------------------------- */
 
   /**
    * Roll a d10 to determine who goes first. On a 6+, heroes do.
@@ -28,6 +33,8 @@ export default class DrawSteelCombat extends foundry.documents.Combat {
     }, { rollMode: CONST.DICE_ROLL_MODES.PUBLIC });
   }
 
+  /* -------------------------------------------------- */
+
   /** @inheritdoc */
   async startCombat() {
     for (const combatant of this.combatants) {
@@ -42,6 +49,8 @@ export default class DrawSteelCombat extends foundry.documents.Combat {
     return super.startCombat();
   }
 
+  /* -------------------------------------------------- */
+
   /**
    * @inheritdoc In Draw Steel's default initiative, non-GM users cannot change the round
    * @param {User} user The user attempting to change the round
@@ -51,6 +60,8 @@ export default class DrawSteelCombat extends foundry.documents.Combat {
     if (game.settings.get(systemID, "initiativeMode") !== "default") return super._canChangeRound(user);
     return user.isGM;
   }
+
+  /* -------------------------------------------------- */
 
   /** @inheritdoc */
   async nextRound() {
@@ -62,6 +73,8 @@ export default class DrawSteelCombat extends foundry.documents.Combat {
     const combatantUpdates = this.combatants.map(c => ({ _id: c.id, initiative: c.actor?.system.combat.turns ?? 1 }));
     this.updateEmbeddedDocuments("Combatant", combatantUpdates);
   }
+
+  /* -------------------------------------------------- */
 
   /** @inheritdoc */
   async endCombat() {
@@ -77,6 +90,8 @@ export default class DrawSteelCombat extends foundry.documents.Combat {
 
     return deletedCombat;
   }
+
+  /* -------------------------------------------------- */
 
   /**
    * @param {DrawSteelCombatant | DrawSteelCombatantGroup} a Some combatant
@@ -101,6 +116,28 @@ export default class DrawSteelCombat extends foundry.documents.Combat {
     return super._sortCombatants(a, b);
   }
 
+  /* -------------------------------------------------- */
+
+  /** @inheritdoc */
+  _onDelete(options, userId) {
+    super._onDelete(options, userId);
+    if (!game.user.isActiveGM) return;
+
+    for (const combatant of this.combatants) {
+      const actor = combatant.actor;
+      if (!actor) continue;
+      /** @type {ActiveEffectData[]} */
+      const updates = [];
+      for (const effect of actor.appliedEffects) {
+        if (!(effect.system instanceof BaseEffectModel)) continue;
+        if (effect.system.end.type === "encounter") updates.push({ _id: effect.id, disabled: true });
+      }
+      actor.updateEmbeddedDocuments("ActiveEffect", updates);
+    }
+  }
+
+  /* -------------------------------------------------- */
+
   /**
      * Actions taken after descendant documents have been created and changes have been applied to client data.
      * @param {DrawSteelCombat} parent         The direct parent of the created Documents, may be this Document or a child
@@ -116,6 +153,8 @@ export default class DrawSteelCombat extends foundry.documents.Combat {
     super._onCreateDescendantDocuments(parent, collection, documents, data, options, userId);
     if (collection === "groups") this.#onModifyCombatantGroups(parent, documents, options);
   }
+
+  /* -------------------------------------------------- */
 
   /**
      * Actions taken after descendant documents have been updated and changes have been applied to client data.
@@ -133,6 +172,8 @@ export default class DrawSteelCombat extends foundry.documents.Combat {
     if (collection === "groups") this.#onModifyCombatantGroups(parent, documents, options);
   }
 
+  /* -------------------------------------------------- */
+
   /**
      * Actions taken after descendant documents have been deleted and those deletions have been applied to client data.
      * @param {Document} parent         The direct parent of the deleted Documents, may be this Document or a child
@@ -149,6 +190,8 @@ export default class DrawSteelCombat extends foundry.documents.Combat {
     if (collection === "groups") this.#onModifyCombatantGroups(parent, documents, options);
   }
 
+  /* -------------------------------------------------- */
+
   /**
    * Shared actions taken when CombatantGroups are modified within this Combat document.
    * @param {DrawSteelCombat} parent              The direct parent of the created Documents, may be this Document or a child
@@ -159,6 +202,32 @@ export default class DrawSteelCombat extends foundry.documents.Combat {
     if ((ui.combat.viewed === parent) && (options.render !== false)) ui.combat.render();
   }
 
+  /**
+   * Handle Draw Steel effect expiration logic
+   * @inheritdoc
+   */
+  async _onEndTurn(combatant, context) {
+    /** @type {DrawSteelActor} */
+    const actor = combatant.actor;
+    if (!actor) return;
+    /** @type {ActiveEffectData[]} */
+    const updates = [];
+    for (const effect of actor.appliedEffects) {
+      if (!(effect.system instanceof BaseEffectModel)) continue;
+      switch (effect.system.end.type) {
+        case "turn":
+          updates.push({ _id: effect.id, disabled: true });
+          break;
+        case "save":
+          SavingThrowManager.delegateSavingThrow(effect);
+          break;
+      }
+    }
+    actor.updateEmbeddedDocuments("ActiveEffect", updates);
+  }
+
+  /* -------------------------------------------------- */
+
   /** @inheritdoc */
   async _onStartRound() {
     /** @type {MaliceModel} */
@@ -168,6 +237,8 @@ export default class DrawSteelCombat extends foundry.documents.Combat {
       .map(c => c.actor);
     await malice._onStartRound(this, aliveHeroes);
   }
+
+  /* -------------------------------------------------- */
 
   /**
    * Prompt the GM for a number of victories to award and then award each character that many victories