+++ CHANGELOG.md
@@ -20,6 +20,13 @@
 
 ### 0.7.2
 
+### Added
+- Added support for the automatic end of effects
+  - End of Turn will automatically self-disable when an affected actor ends their turn.
+  - Saving Throws will create prompts for owners. If multiple players own an actor, the active GM will receive a dialog to help delegate rolls.
+    - The roll message has a button to spend a hero token to automatically succeed
+  - End of Encounter effects will automatically self-disable alongside the encounter.
+
 ### Changed
 - Aligned the system's active effect suppression checks with core v13 behavior, allowing the system model to also apply.